Fix bug for classes with non-graphql attribute classes
I have a class with a non-graphql attribute class (`<<__ConsistentConstruct>>`) that was being picked up at this line and leading to codegen failures. the code expects my non-graphql class to exist in `$class_fields`, and this PR simply continues if the class has attribute classes but no graphql attribute classes.  another solution could be to more narrowly tailor the directories that my codegen script checks, but seems like we should be able to handle this case fairly well.